loongarch: don't throw ABI/multilib errors at all
Again, no upstreamed ports do this.  All of these ports just let the
linker to report an error.

Checking this in GCC driver is too paranoid.  It prevents building GRUB
using a non-multilib GCC with no real reason (the soft-float part of GRUB
is using -nostdlib, so it does not require multilib at all).

gcc/

	* config/loongarch/loongarch-driver.c
	  (driver_get_normalized_m_opts): Don't check multilib too
	  paranoidly.